updated the pipe and pipe mapper tests.
