+++ unit_test/adapters/transforms/esdl_asset_mappers/test_esdl_asset_pipe_mapper.py
@@ -13,24 +13,27 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-"""Test Junction entities."""
+"""Test pipe mapper."""
 import unittest
 from pathlib import Path
+from unittest.mock import Mock
+
 from omotes_simulator_core.adapter.transforms.esdl_asset_mappers.pipe_mapper import (
     EsdlAssetPipeMapper,
 )
 from omotes_simulator_core.entities.esdl_object import EsdlObject
 from omotes_simulator_core.infrastructure.utils import pyesdl_from_file
 from omotes_simulator_core.entities.assets.asset_defaults import PIPE_DEFAULTS
+from omotes_simulator_core.entities.assets.pipe import Pipe
+from omotes_simulator_core.entities.assets.esdl_asset_object import EsdlAssetObject
 
 
 class TestEsdlAssetPipeMapper(unittest.TestCase):
+    """Test class for pipe mapper."""
 
     def setUp(self):
         """Define the variables used in the tests."""
-        self.name: str = "pipe"
-        self.asset_id: str = "pipe_id"
-        self.ports: list[str] = ["test1", "test2"]
+
         self.length: float = 1
         self.inner_diameter: float = 1
         self.roughness: float = 0.001
@@ -39,7 +42,7 @@ class TestEsdlAssetPipeMapper(unittest.TestCase):
         self.external_temperature: float = 273.15 + 20.0
         self.qheat_external: float = 0.0
 
-        """Arranging standard objects for the other test."""
+        # Arranging standard objects for the other test
         esdl_file_path = (
             Path(__file__).parent / ".." / ".." / ".." / ".." / "testdata" / "test1.esdl"
         )
@@ -47,14 +50,12 @@ class TestEsdlAssetPipeMapper(unittest.TestCase):
         self.esdl_object = EsdlObject(pyesdl_from_file(esdl_file_path))
         self.mapper = EsdlAssetPipeMapper()
 
-    def test_get_all_assets_of_type(self):
+    def test_to_entity_method(self):
+        """Test for to_entity method."""
         pipes = self.esdl_object.get_all_assets_of_type("pipe")  # act
         self.assertGreater(len(pipes), 0, "No pipes found in the ESDL file.")
-        self.pipe = pipes[0]
-
-    def test_to_entity(self):
-        self.test_get_all_assets_of_type()
-        esdl_asset = self.pipe
+        self.esdl_asset = pipes[0]
+        esdl_asset = pipes[0]
         pipe_entity = self.mapper.to_entity(esdl_asset)
         self.assertEqual(pipe_entity.name, esdl_asset.esdl_asset.name)
         self.assertEqual(pipe_entity.asset_id, esdl_asset.esdl_asset.id)
@@ -63,9 +64,68 @@ class TestEsdlAssetPipeMapper(unittest.TestCase):
         self.assertEqual(pipe_entity.diameter, self.mapper._get_diameter(esdl_asset))  # type: ignore
         self.assertEqual(pipe_entity.roughness, esdl_asset.get_property("roughness", PIPE_DEFAULTS.roughness)[0])  # type: ignore
         self.assertEqual(pipe_entity.alpha_value, self.mapper._get_heat_transfer_coefficient(esdl_asset))  # type: ignore
-        self.assertEqual(pipe_entity.minor_loss_coefficient, esdl_asset.get_property("minor_loss_coefficient", PIPE_DEFAULTS.minor_loss_coefficient)[0])  # type: ignore
-        self.assertEqual(pipe_entity.external_temperature, esdl_asset.get_property("external_temperature", PIPE_DEFAULTS.external_temperature)[0])  # type: ignore
+        self.assertEqual(
+            pipe_entity.minor_loss_coefficient,  # type: ignore
+            esdl_asset.get_property("minor_loss_coefficient", PIPE_DEFAULTS.minor_loss_coefficient)[
+                0
+            ],
+        )
+        self.assertEqual(
+            pipe_entity.external_temperature,  # type: ignore
+            esdl_asset.get_property("external_temperature", PIPE_DEFAULTS.external_temperature)[0],
+        )
         self.assertEqual(
             pipe_entity.qheat_external,  # type: ignore
             esdl_asset.get_property("qheat_external", PIPE_DEFAULTS.qheat_external)[0],
         )
+
+    def test_pipe_get_property_diameter(self):
+        """Evaluate the get property diameter method to retrieve diameters."""
+        # Arrange
+        esdl_asset_mock = Mock()
+        esdl_asset_mock.get_property.return_value = (1.0, True)
+
+        # Act
+        pass
+
+        # Assert
+        self.assertEqual(EsdlAssetPipeMapper._get_diameter(self, esdl_asset=esdl_asset_mock), 1.0)
+
+    def test_pipe_get_property_diameter_failed(self):
+        """Evaluate failure to retrieve diameter from ESDL asset."""
+        # Arrange
+        esdl_asset_mock = Mock()
+        esdl_asset_mock.get_property.return_value = (1.0, False)
+
+        # Act
+        pass
+
+        # Assert
+        with self.assertRaises(NotImplementedError):
+            EsdlAssetPipeMapper._get_diameter(self, esdl_asset=esdl_asset_mock)
+
+    def test_pipe_get_heat_transfer_coefficient(self):
+        """Evaluate the get heat transfer coefficient method."""
+        # Arrange
+        # - Load esdl pipe asset
+        esdl_file_path = (
+            Path(__file__).parent
+            / ".."
+            / ".."
+            / ".."
+            / ".."
+            / "testdata"
+            / "test_pipe_material.esdl"
+        )
+        esdl_file_path = str(esdl_file_path)
+        esdl_object = EsdlObject(pyesdl_from_file(esdl_file_path))
+        esdl_pipes = esdl_object.get_all_assets_of_type("pipe")
+        esdl_pipe = [pipe for pipe in esdl_pipes if pipe.esdl_asset.name == "pipe_with_material"][0]
+
+        # Act
+        alpha_value = EsdlAssetPipeMapper._get_heat_transfer_coefficient(
+            self, esdl_asset=esdl_pipe
+        )  # act
+
+        # Assert
+        self.assertEqual(alpha_value, 0.8901927763663371)
